refactor: Extract tag filter parsing & matching logic to separate class
A lot of our tests for TagFilter are around the specifics of parsing and
matching different types of legacy filter expression. A lot of variants
are then either:

* Repeated for matching against feature, feature & scenario, feature &
  scenario & table union tags
* Tested only on one level of the hierarchy with a comment that other
  variants are not tested because we know the tag matching logic is
  shared and we assume those expressions have been covered at another
  level.

This makes the TagFilter tests themselves hard to follow or refactor,
without a lot of duplication.

Moving the specifics of matching a filter string to a set of tags to a
standalone class will then allow us to move the edge case etc tests,
so that the TagFilter tests can cover a smaller set of proving
behaviour - in particular around what gets included when filtering a
feature.
