+++ src/Filter/TagFilter.php
@@ -26,16 +26,11 @@ class TagFilter extends ComplexFilter
      */
     protected $filterString;
 
-    /**
-     * @var array{all?: list<array{any: list<array{tag: string, hasTag: bool}>}>, filterString: string}
-     */
-    private array $parsedFilter;
+    private TagFilterMatcher $filterMatcher;
 
     public function __construct(string $filterString)
     {
-        $this->filterString = trim($filterString);
-        $this->parseFilterString();
-
+        $this->filterMatcher = new TagFilterMatcher($filterString);
         // Because `filterString` is protected (and therefore could in theory be modified by a child class at runtime),
         // we need to check if the parsed filter is up to date every time isTagsMatchCondition is called.
         //
@@ -45,83 +40,7 @@ class TagFilter extends ComplexFilter
         //
         // This can all be removed in the next major if we make `filterString` private and/or readonly and remove the
         // normalisation of deprecated syntax.
-        $this->filterString = $this->parsedFilter['filterString'] = implode(
-            '&&',
-            array_map(
-                static fn (array $filterClause) => implode(',',
-                    array_map(
-                        static fn (array $tag) => sprintf(
-                            '%s%s',
-                            $tag['hasTag'] ? '' : '~',
-                            $tag['tag']
-                        ),
-                        $filterClause['any']
-                    )),
-                $this->parsedFilter['all'] ?? [],
-            ),
-        );
-    }
-
-    private function parseFilterString(): void
-    {
-        $this->parsedFilter = [
-            'filterString' => $this->filterString,
-        ];
-
-        if ($this->filterString === '') {
-            return;
-        }
-
-        $hadTagWithWhitespace = false;
-        $hadTagWithoutPrefix = false;
-
-        $this->parsedFilter['all'] = [];
-        foreach (explode('&&', $this->filterString) as $andTags) {
-            $orParts = [];
-            foreach (explode(',', $andTags) as $tag) {
-                $tag = trim($tag);
-
-                // Fix tag expressions where the filter string does not include the `@` prefixes.
-                // e.g. `new TagFilter('wip&&~slow')` rather than `new TagFilter('@wip&&~@slow')`. These were
-                // historically supported, although not officially, and have been reinstated to solve a BC issue.
-                // This syntax is deprecated and will be removed in future.
-                $fixedTag = match (true) {
-                    // Valid - tag filter contains the `@` prefix
-                    str_starts_with($tag, '@'),
-                    str_starts_with($tag, '~@'),
-                    // Valid historical edge case - tag filter contains the `@` prefix, but there is whitespace after the `~`
-                    (bool) preg_match('/^~\s+@/', $tag) => $tag,
-                    // Invalid / legacy cases - insert the missing `@` prefix in the right place
-                    str_starts_with($tag, '~') => '~@' . substr($tag, 1),
-                    default => '@' . $tag,
-                };
-
-                if (str_starts_with($fixedTag, '~')) {
-                    $orParts[] = ['tag' => substr($fixedTag, 1), 'hasTag' => false];
-                } else {
-                    $orParts[] = ['tag' => $fixedTag, 'hasTag' => true];
-                }
-
-                $hadTagWithoutPrefix = $hadTagWithoutPrefix || ($tag !== $fixedTag);
-                $hadTagWithWhitespace = $hadTagWithWhitespace || str_contains($tag, ' ');
-            }
-
-            $this->parsedFilter['all'][] = ['any' => $orParts];
-        }
-
-        if ($hadTagWithWhitespace) {
-            trigger_error(
-                'Tags with whitespace are deprecated and may be removed in a future version',
-                E_USER_DEPRECATED
-            );
-        }
-
-        if ($hadTagWithoutPrefix) {
-            trigger_error(
-                'Filter strings should contain `@` prefixes for tags, e.g. `@wip` rather than `wip`.',
-                E_USER_DEPRECATED
-            );
-        }
+        $this->filterString = $this->filterMatcher->getNormalisedFilterString();
     }
 
     /**
@@ -199,38 +118,12 @@ class TagFilter extends ComplexFilter
      */
     protected function isTagsMatchCondition(array $tags)
     {
-        if ($this->parsedFilter['filterString'] !== $this->filterString) {
+        if ($this->filterMatcher->getNormalisedFilterString() !== $this->filterString) {
             // A child class has modified the filter string since the last call.
-            $this->parseFilterString();
+            $this->filterMatcher = new TagFilterMatcher($this->filterString);
+            $this->filterString = $this->filterMatcher->getNormalisedFilterString();
         }
 
-        if (!isset($this->parsedFilter['all'])) {
-            return true;
-        }
-
-        // If the file was parsed in legacy mode, the `@` prefix will have been removed from the individual tags on the
-        // parsed node. The tags in the filter expression still have their @ so we add the prefix back here if required.
-        // This can be removed once legacy parsing mode is removed.
-        $tags = array_map(
-            static fn (string $tag) => str_starts_with($tag, '@') ? $tag : '@' . $tag,
-            $tags
-        );
-
-        foreach ($this->parsedFilter['all'] as $filterPart) {
-            $satisfiesComma = false;
-
-            foreach ($filterPart['any'] as $tag) {
-                if (in_array($tag['tag'], $tags, true) === $tag['hasTag']) {
-                    $satisfiesComma = true;
-                    break;
-                }
-            }
-
-            if (!$satisfiesComma) {
-                return false;
-            }
-        }
-
-        return true;
+        return $this->filterMatcher->isMatch($tags);
     }
 }
